refactor(hash provider): move HashProvider to shared dir
