--- src/modules/user/providers/index.ts
@@ -1,6 +1,0 @@
-import { container } from 'tsyringe';
-
-import { IHashProvider } from './HashProvider/models/IHashProvider';
-import { BcryptHashProvider } from './HashProvider/implementations/BcryptHashProvider';
-
-container.registerSingleton<IHashProvider>('HashProvider', BcryptHashProvider);
