+++ src/shared/container/index.ts
@@ -1,7 +1,7 @@
 import './providers';
 import '@modules/specification';
 import '@modules/category';
+import '@modules/user';
 import '@modules/car';
 import '@modules/carImage';
-import '@modules/user';
 import '@modules/rental';
