+++ src/shared/container/providers/HashProvider/dtos/ICompareHashDTO.ts
