+++ src/shared/container/providers/HashProvider/index.ts
@@ -1,0 +1,6 @@
+import { container } from 'tsyringe';
+
+import { IHashProvider } from './models/IHashProvider';
+import { BcryptHashProvider } from './implementations/BcryptHashProvider';
+
+container.registerSingleton<IHashProvider>('HashProvider', BcryptHashProvider);
