+++ src/shared/container/providers/HashProvider/models/IHashProvider.ts
