+++ src/shared/container/providers/HashProvider/tests/HashProvider.spec.ts
