+++ src/shared/container/providers/index.ts
@@ -1,1 +1,2 @@
 import './DateProvider';
+import './HashProvider';
